feat(system): 添加对象存储配置状态修改功能并优化相关逻辑
- 在 SysOssConfigController 中添加 changeStatus 方法，用于修改配置状态
- 在 SysOssConfigService 接口中添加 updateOssConfigStatus 方法
- 在 SysOssConfigServiceImpl 中实现 updateOssConfigStatus 方法，包括缓存更新和事务处理
- 优化 SysOssConfigSaveReq 和 SysOssConfigUpdateReq 中的字段验证逻辑
- 添加 configKey 唯一性校验逻辑
-修复删除配置时未更新缓存的问题

+++ aixbox-modules/aixbox-system/src/main/java/com/aixbox/system/service/impl/SysOssConfigServiceImpl.java
@@ -1,8 +1,22 @@
 package com.aixbox.system.service.impl;
 
+import cn.hutool.core.collection.CollUtil;
+import cn.hutool.core.util.ObjectUtil;
+import com.aixbox.common.core.constant.CacheNames;
+import com.aixbox.common.core.exception.ServiceException;
 import com.aixbox.common.core.pojo.PageResult;
+import com.aixbox.common.core.utils.StrUtils;
+import com.aixbox.common.core.utils.json.JsonUtils;
 import com.aixbox.common.core.utils.object.BeanUtils;
+import com.aixbox.common.core.utils.object.MapstructUtils;
+import com.aixbox.common.core.utils.object.ObjectUtils;
+import com.aixbox.common.oss.constant.OssConstant;
+import com.aixbox.common.redis.utils.CacheUtils;
+import com.aixbox.common.redis.utils.RedisUtils;
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
 import com.aixbox.system.domain.entity.SysOssConfig;
 import com.aixbox.system.domain.vo.request.SysOssConfigPageReq;
@@ -10,6 +24,7 @@ import com.aixbox.system.domain.vo.request.SysOssConfigSaveReq;
 import com.aixbox.system.domain.vo.request.SysOssConfigUpdateReq;
 import com.aixbox.system.mapper.SysOssConfigMapper;
 import com.aixbox.system.service.SysOssConfigService;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -34,7 +49,12 @@ public class SysOssConfigServiceImpl implements SysOssConfigService {
     public Long addSysOssConfig(SysOssConfigSaveReq addReq) {
         SysOssConfig  sysOssConfig = BeanUtils.toBean(addReq, SysOssConfig.class);
         validEntityBeforeSave(sysOssConfig);
-        sysOssConfigMapper.insert(sysOssConfig);
+        boolean flag = sysOssConfigMapper.insert(sysOssConfig) > 0;
+        if (flag) {
+            // 从数据库查询完整的数据做缓存
+            sysOssConfig = sysOssConfigMapper.selectById(sysOssConfig.getId());
+            CacheUtils.put(CacheNames.SYS_OSS_CONFIG, sysOssConfig.getConfigKey(), JsonUtils.toJsonString(sysOssConfig));
+        }
         return sysOssConfig.getId();
     }
 
@@ -47,7 +67,13 @@ public class SysOssConfigServiceImpl implements SysOssConfigService {
     public Boolean updateSysOssConfig(SysOssConfigUpdateReq updateReq) {
         SysOssConfig sysOssConfig = BeanUtils.toBean(updateReq, SysOssConfig.class);
         validEntityBeforeSave(sysOssConfig);
-        return sysOssConfigMapper.updateById(sysOssConfig) > 0;
+        boolean flag = sysOssConfigMapper.updateById(sysOssConfig) > 0;
+        if (flag) {
+            // 从数据库查询完整的数据做缓存
+            sysOssConfig = sysOssConfigMapper.selectById(sysOssConfig.getId());
+            CacheUtils.put(CacheNames.SYS_OSS_CONFIG, sysOssConfig.getConfigKey(), JsonUtils.toJsonString(sysOssConfig));
+        }
+        return flag;
     }
 
     /**
@@ -57,8 +83,21 @@ public class SysOssConfigServiceImpl implements SysOssConfigService {
      */
     @Override
     public Boolean deleteSysOssConfig(List<Long> ids) {
+        if (CollUtil.containsAny(ids, OssConstant.SYSTEM_DATA_IDS)) {
+            throw new ServiceException("系统内置, 不可删除!");
+        }
         validateSysOssConfigExists(ids);
-        return sysOssConfigMapper.deleteByIds(ids) > 0;
+        List<SysOssConfig> list = CollUtil.newArrayList();
+        for (Long configId : ids) {
+            SysOssConfig config = sysOssConfigMapper.selectById(configId);
+            list.add(config);
+        }
+        boolean flag = sysOssConfigMapper.deleteByIds(ids) > 0;
+        if (flag) {
+            list.forEach(sysOssConfig ->
+                    CacheUtils.evict(CacheNames.SYS_OSS_CONFIG, sysOssConfig.getConfigKey()));
+        }
+        return flag;
     }
 
     /**
@@ -93,10 +132,43 @@ public class SysOssConfigServiceImpl implements SysOssConfigService {
     }
 
     /**
+     * 启用禁用状态
+     */
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public int updateOssConfigStatus(SysOssConfigUpdateReq bo) {
+        SysOssConfig sysOssConfig = BeanUtils.toBean(bo, SysOssConfig.class);
+        int row = sysOssConfigMapper.update(null, new LambdaUpdateWrapper<SysOssConfig>()
+                .set(SysOssConfig::getStatus, "1"));
+        row += sysOssConfigMapper.updateById(sysOssConfig);
+        if (row > 0) {
+            RedisUtils.setCacheObject(OssConstant.DEFAULT_CONFIG_KEY, sysOssConfig.getConfigKey());
+        }
+        return row;
+    }
+
+    /**
      * 保存前的数据校验
      */
     private void validEntityBeforeSave(SysOssConfig entity){
-        //TODO 做一些数据校验,如唯一约束
+        if (StrUtils.isNotEmpty(entity.getConfigKey())
+                && !checkConfigKeyUnique(entity)) {
+            throw new ServiceException("操作配置'" + entity.getConfigKey() + "'失败, 配置key已存在!");
+        }
+    }
+
+    /**
+     * 判断configKey是否唯一
+     */
+    private boolean checkConfigKeyUnique(SysOssConfig sysOssConfig) {
+        long ossConfigId = ObjectUtils.notNull(sysOssConfig.getId(), -1L);
+        SysOssConfig info = sysOssConfigMapper.selectOne(new LambdaQueryWrapper<SysOssConfig>()
+                .select(SysOssConfig::getId, SysOssConfig::getConfigKey)
+                .eq(SysOssConfig::getConfigKey, sysOssConfig.getConfigKey()));
+        if (ObjectUtil.isNotNull(info) && info.getId() != ossConfigId) {
+            return false;
+        }
+        return true;
     }
 }
